fix: resolve all remaining clippy warnings
- cpu_local: add Default impl for TlbShootdownEntry/Mailbox/PerCpuData
- cpu_local: allow interior mutability in const LAPIC maps
- sync_safe: remove redundant explicit auto-deref
- drivers/framebuffer: add Default impl, use enumerate, fix range check
- virtio/queue: use div_ceil, use contains() instead of iter().any()
- bootloader: remove unnecessary u64 casts
- build.rs: add allow(dead_code), use strip_prefix, collapse if

Addresses all clippy errors blocking CI

+++ bootloader/src/main.rs
@@ -884,7 +884,7 @@ fn efi_main(handle: Handle, mut system_table: SystemTable<Boot>) -> Status {
         // This is defense-in-depth: even if an attacker can write to page tables via
         // the recursive mapping, they cannot execute code from them.
         (&mut *pml4_ptr)[510].set_addr(
-            PhysAddr::new(pml4_frame as u64),
+            PhysAddr::new(pml4_frame),
             Flags::PRESENT | Flags::WRITABLE | Flags::NO_EXECUTE,
         );
 
@@ -898,7 +898,7 @@ fn efi_main(handle: Handle, mut system_table: SystemTable<Boot>) -> Status {
 
         // 加载新的页表
         Cr3::write(
-            PhysFrame::containing_address(PhysAddr::new(pml4_frame as u64)),
+            PhysFrame::containing_address(PhysAddr::new(pml4_frame)),
             Cr3::read().1,
         );
